+++ example/lib/pages/input_otp.dart
@@ -67,7 +67,7 @@ class _InputOTPPageState extends State<InputOTPPage> {
                   ShadInputOTPSlot(),
                 ],
               ),
-              ShadImage.square(size: 24, LucideIcons.dot),
+              Icon(LucideIcons.dot),
               ShadInputOTPGroup(
                 children: [
                   ShadInputOTPSlot(),