+++ lib/shadcn_ui.dart
@@ -26,7 +26,6 @@ export 'src/components/form/fields/switch.dart';
 export 'src/components/form/fields/time_picker.dart';
 export 'src/components/form/fields/input_otp.dart';
 export 'src/components/form/form.dart';
-export 'src/components/image.dart';
 export 'src/components/input.dart';
 export 'src/components/input_otp.dart';
 export 'src/components/popover.dart';