+++ lib/src/components/form/fields/date_picker.dart
@@ -7,7 +7,6 @@ import 'package:shadcn_ui/src/components/button.dart';
 import 'package:shadcn_ui/src/components/calendar.dart';
 import 'package:shadcn_ui/src/components/date_picker.dart';
 import 'package:shadcn_ui/src/components/form/field.dart';
-import 'package:shadcn_ui/src/components/image.dart';
 import 'package:shadcn_ui/src/components/popover.dart';
 import 'package:shadcn_ui/src/raw_components/portal.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
@@ -151,11 +150,11 @@ class ShadDatePickerFormField extends ShadFormBuilderField<DateTime> {
     /// {@macro ShadCalendar.navigationButtonIconSize}
     double? navigationButtonIconSize,
 
-    /// {@macro ShadCalendar.backNavigationButtonSrc}
-    ShadImageSrc? backNavigationButtonSrc,
+    /// {@macro ShadCalendar.backNavigationButtonIconData}
+    IconData? backNavigationButtonIconData,
 
-    /// {@macro ShadCalendar.forwardNavigationButtonSrc}
-    ShadImageSrc? forwardNavigationButtonSrc,
+    /// {@macro ShadCalendar.forwardNavigationButtonIconData}
+    IconData? forwardNavigationButtonIconData,
 
     /// {@macro ShadCalendar.navigationButtonPadding}
     EdgeInsets? navigationButtonPadding,
@@ -288,8 +287,8 @@ class ShadDatePickerFormField extends ShadFormBuilderField<DateTime> {
     /// {@macro ShadButton.icon}
     Widget? icon,
 
-    /// {@macro ShadDatePicker.iconSrc}
-    ShadImageSrc? iconSrc,
+    /// {@macro ShadDatePicker.iconData}
+    IconData? iconData,
 
     /// {@macro ShadButton.child}
     Widget? buttonChild,
@@ -472,8 +471,8 @@ class ShadDatePickerFormField extends ShadFormBuilderField<DateTime> {
               monthSelectorPadding: monthSelectorPadding,
               navigationButtonSize: navigationButtonSize,
               navigationButtonIconSize: navigationButtonIconSize,
-              backNavigationButtonSrc: backNavigationButtonSrc,
-              forwardNavigationButtonSrc: forwardNavigationButtonSrc,
+              backNavigationButtonIconData: backNavigationButtonIconData,
+              forwardNavigationButtonIconData: forwardNavigationButtonIconData,
               navigationButtonPadding: navigationButtonPadding,
               navigationButtonDisabledOpacity: navigationButtonDisabledOpacity,
               spacingBetweenMonths: spacingBetweenMonths,
@@ -515,7 +514,7 @@ class ShadDatePickerFormField extends ShadFormBuilderField<DateTime> {
               useSameGroupIdForChild: useSameGroupIdForChild,
               onPressed: onPressed,
               onLongPress: onLongPress,
-              iconSrc: iconSrc,
+              iconData: iconData,
               buttonChild: buttonChild,
               buttonVariant: buttonVariant,
               size: size,